refactor(Restaurant): 식당 템플릿 분리
