+++ src/index.js
@@ -2,8 +2,8 @@ import Restaurants from './domains/Restaurants';
 
 import Header from './components/Header/Header';
 import RestaurantCreationModal from './components/AddRestaurantModal/AddRestaurantModal';
-import Restaurant from './components/Restaurant/Restaurant';
 import RestaurantFilter from './components/RestaurantFilter/RestaurantFilter';
+import Restaurant from './components/Common/Restaurant/Restaurant';
 
 import { $ } from './utils/dom';
 
@@ -16,13 +16,12 @@ document.addEventListener('DOMContentLoaded', () => {
   // components
   const header = new Header();
   const restaurantFilter = new RestaurantFilter(restaurants);
-  const restaurant = new Restaurant();
   const modal = new RestaurantCreationModal(restaurants);
 
   $('header').innerHTML = header.render();
   $('restaurant-filter-container').innerHTML += restaurantFilter.render();
   restaurants.standardList.forEach((restaurantData) => {
-    $('restaurant-list').innerHTML += restaurant.render(restaurantData);
+    $('restaurant-list').innerHTML += Restaurant(restaurantData);
   });
   $('add-restaurant-modal').innerHTML = modal.render();
 });
